cat_dog_classification/app/preprocess_service/requirements.txt
tensorflow==2.5.0
tensorflow-serving-api==2.5.0
numpy
scipy
Pillow==8.0.1
opencv-python==4.5.3.56
uvicorn
fastapi